them, sua, xoa san pham okey co ban

+++ app/Http/Controllers/CategoryController.php
@@ -15,76 +15,81 @@ use Mail;
 class CategoryController extends Controller
 {
 
-
     public function __construct()
     {
-        $title = 'Dashboard - Category';
-        $class_name = substr(__CLASS__, 21);
+        /**
+         * SET title, class_name, action_name
+         */
+        $title       = 'Dashboard - Category';
+        $class_name  = substr(__CLASS__, 21);
         $action_name = substr(strrchr(Route::currentRouteAction(), "@"), 1);
         View::share(array(
-            'title' => $title,
-            'class_name' => $class_name,
+            'title'       => $title,
+            'class_name'  => $class_name,
             'action_name' => $action_name,
         ));
-        $this->afterFilter(function () {
-            // something
-        });
+
     }
 
     public function index()
     {
         $categories = Category::all()->toArray();
+        $result     = null;
+        getAllCategory($categories, $parent = 0, $text = "", $select = 0, $result);
 
-        return View::make('category.list', ['categories' => $categories]);
+        return View::make('category.list', ['categories' => $result]);
     }
 
     public function getJsonData()
     {
         $categories = Category::all();
-//        foreach($categories as $category){
-//            $categories[] = Category::where('parent', '=', $category['id'])->get();
-//         }
-//        dd($categories->toArray());
-        //dd($categories->toArray());
         foreach ($categories as $category) {
             if ($category['parent'] == 0) {
                 $category['parent'] = "#";
             }
-            $newCategoryConvert[] = array('id' => $category['id'],
-                'parent' => $category['parent'],
-                'text' => $category['category_name'],
+            $newCategoryConvert[] = array('id'     => $category['id'],
+                                          'parent' => $category['parent'],
+                                          'text'   => $category['category_name'],
             );
 
         }
 
         return $newCategoryConvert;
     }
+
     public function create()
     {
         $categories = Category::all()->toArray();
+        $result     = null;
+        getAllCategory($categories, $parent = 0, $text = "", $select = 0, $result);
 
         return view('category.create')->with([
-            "categories" => $categories,
+            "categories" => $result,
 
         ]);
 
     }
+
     public function update(CategoryRequest $request)
     {
         $allRequest = $request->all();
-        $model = Category::find($allRequest['id']);
+        $model      = Category::find($allRequest['id']);
 
-        autoAssignDataToProperty($model,$request->all());
+        autoAssignDataToProperty($model, $request->all());
         $model->save();
         return redirect()->action('CategoryController@index')->withSuccess(Lang::get('messages.update_success'));
 
 
     }
+
     public function delete(Request $request)
     {
         $allRequest = $request->all();
-        $model = Category::find($allRequest['id']);
-        if(Category::where('parent', $model->id)->count() !=0){
+        $model      = Category::find($allRequest['id']);
+        if ($model == null) {
+            return redirect()->action('CategoryController@index')->withErrors(Lang::get('messages.no_id'));
+        }
+        if (Category::where('parent', $model->id)->count() != 0) {
             return redirect()->action('CategoryController@index')->withErrors(Lang::get('messages.has_childrent'));
         }
         $model->delete();
@@ -95,20 +100,20 @@ class CategoryController extends Controller
 
     public function store(CategoryRequest $request)
     {
-        $model=new Category();
-        autoAssignDataToProperty($model,$request->all());
+        $model = new Category();
+        autoAssignDataToProperty($model, $request->all());
         $model->save();
         return redirect()->action('CategoryController@index')
             ->withSuccess(Lang::get('messages.create_success'));
     }
 
-    public function test()
-    {
-        $categories = Category::all()->toArray();
-        $data=array();
-        getAllCategoryTest($categories,$data);
-        dd($data);
-    }
+//    public function test()
+//    {
+//        $categories = Category::all()->toArray();
+//        $data       = array();
+//        getAllCategoryTest($categories, $data);
+//        dd($data);
+//    }
 
 }
